add new way to measure BF and (1-Q)/Q ratio
In this commit, I add in addition to post prob, the tree vote probability in barrier detection. Furthermore, the traditional way to compute the average (1-Q)/Q might be biased by cases where Q=0. In consequences, I added an approximate computation of the ratio which is less sensitive to cases where Q=0.